myocamlbuild.ml: use correct globbing patterns
The previous globbing patterns of the form moc_%.c couldn't work if
the file was in a subdirectory such as src/moc_foo.c. You have to use
instead something of the form %moc_%.c, and then name the patterns to
avoid ambiguities (as examples can be found in
ocamlbuild/ocaml_specific.ml in the OCaml distribution).